refactor fill
