+++ package.json
@@ -39,6 +39,7 @@
     "openai": "^4.3.1",
     "postcss": "8.4.28",
     "posthog-js": "^1.77.2",
+    "promptlayer": "^0.0.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tailwindcss": "3.3.3",
