+++ src/components/landing/Team.tsx
@@ -15,7 +15,6 @@ const Team: React.FC<Props> = ({ members }) => {
   return (
     <div className="container mx-auto p-8">
       <div className="">
-        {/* <h2 className="text-2xl block text-center  font-semibold mb-4">About the Team</h2> */}
         <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
           {members.map((member, index) => (
             <div
