+++ src/components/settings/DashboardMode.tsx
@@ -1,5 +1,5 @@
-import { SpinnerFullPage } from '@/components/home/Spinner';
-import { Form, Response, User } from '@/types';
+import { SpinnerFullPage } from '@/components/settings/Spinner';
+import { Form, Response, User } from '@/models';
 import { getFormsFromSupabase, getResponsesFromSupabase } from '@/utils';
 import { LinkIcon } from '@heroicons/react/24/outline';
 import { createClientComponentClient } from '@supabase/auth-helpers-nextjs';
@@ -152,7 +152,7 @@ export default function DashboardMode(props: { user: User | null }) {
                         </button>
                       </Link>
                       {f.is_open ? (
-                        <Link href={'/forms/fill/' + f.id}>
+                        <Link href={'/forms/entry/' + f.id}>
                           <button className="rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus:outline-none focus-visible:ring-2 focus-visible:ring-indigo-600">
                             <LinkIcon className="h-5 w-5" aria-hidden="true" />
                           </button>
