+++ src/components/settings/ErrorMode.tsx
