+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import NavBar from '@/components/home/NavBar';
+import NavBar from '@/components/settings/NavBar';
 import {
   ChatHistory,
   FloatingTextBox,
@@ -59,7 +59,7 @@ function GetStartedButtons() {
           View on GitHub
         </Link>
         <Link
-          href="/forms/fill/5771953d-a003-4969-9071-fcfff4c5bb10"
+          href="/forms/entry/5771953d-a003-4969-9071-fcfff4c5bb10"
           className={`inline-block text-gray-500 border border-2 border-gray-500 font-semibold py-2 px-4 my-2 mx-1 rounded-lg transition duration-200 ease-in-out hover:bg-gray-200`}
         >
           Fill a sample form
